Static analysis reorg.
